git-svn-id: https://zenmagick.svn.sourceforge.net/svnroot/zenmagick/trunk@41 2e913b60-282b-11dc-bd19-815372732833

+++ zen-cart/admin/zmCleanup.php
@@ -35,12 +35,14 @@ require_once('../zenmagick/admin_init.php');
     array_push($_zm_obsolete_files, DIR_FS_CATALOG . "zenmagick/themes/default/controller");
 
     // delete
-    foreach ($_POST['obsolete'] as $file) {
-      if (is_file($file)) {
-        unlink($file);
-      } else if (is_dir($file)) {
-        rmdir($file);
-      }
+    if (isset($_POST) && array_key_exists('obsolete', $_POST)) {
+        foreach ($_POST['obsolete'] as $file) {
+            if (is_file($file)) {
+                unlink($file);
+            } else if (is_dir($file)) {
+                rmdir($file);
+            }
+        }
     }
 
 ?>
